requirements.txt
configparser
RPi.GPIO